+++ src/main/java/kitchenpos/table/domain/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.table.domain;
 
 import kitchenpos.table.domain.OrderTable;
