+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.table.dto;
 
 import kitchenpos.table.dto.OrderTableRequest;
